requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google.genai
google
GitPython==3.1.40
werkzeug==2.3.7